+++ src/strigui/playground.clj
@@ -24,7 +24,7 @@
                            (assoc-in ["click" :args :x] 100)
                            (assoc-in ["click" :args :y] 400)))))
 
-(defmethod wdg/widget-global-event :mouse-clicked [_ canvas widgets x y]
+(defmethod wdg/widget-global-event :mouse-clicked [_ widgets x y]
   (println "Mouse clicked x:" x " y:" y)
   widgets)